Parsing JSON for turns started
1. outsourced convertJSONToDeviceList to two new methods:
- extractValueFromJSONField
- reduceJSONSourceToElements
These methods have been created to be general and to be reused 
in parsing the other type of JSON (for device settings for each turn)
This step also made both methods cleaner, easier to understand.
2. fetched new method setTurnForEachDevice 
(called 'setRound' in the Delphi form project): 
instead of  using .src-files we are going to standardize the communication with C# through JSON-format here as well.
This method is going to be elaborated further.

The description for different kinds of sound frequencies got also separated in a new file 'sound_frequencies.txt'