setTurnForEachDevice written with json_source
Integrated JSONParsing into method setTurnForEachDevice, according to my theory from now on what setRound was
is callable and working with JSON, meaning it could be used to grant parameters to devices in a standardized way.
Added constants for new error codes. Via error codes we can throw typized (custom) exceptions in C# therefore it should be easier to debug the DLL in the future.