XML-serialization done
We can decide in what format we want dev485 to be transported to C# from Delphi: JSON or XML
Device settings are still serialized and sent in JSON, that is going to be the standard.
TODO: When user clicks btnFelmeres -> stops sending signals to device when using XML-serialization, without clicking btnFelmeres, both scenarios work as expected.